# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/period-locks": {
      "get": {
        "description": "Returns the months that are closed, oldest first, optionally within a range.\n\nNot paginated. An account has at most a handful of locked months per year, and a client needs all of them at once to mark the right rows read only in a report.\n\nThe absence of a month from this list is what \"open\" means. There is no unlocked record.",
        "operationId": "PeriodLocksController_list_v1",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "First month of the range, inclusive.",
            "schema": {
              "example": "2026-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Last month of the range, inclusive.",
            "schema": {
              "example": "2026-12",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The closed periods.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PeriodLockResponseDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "The access token is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List closed periods",
        "tags": [
          "Period locks"
        ]
      },
      "post": {
        "description": "Closes months so their plans and expenses become read only.\n\nSupply either `months` or `quarter`. A quarter expands to its three calendar months, so the stored shape is always one record per month and a single month of a closed quarter can be reopened without a special case.\n\nClosing an already closed month is harmless: it is skipped and its original time is preserved, so locking a quarter that overlaps a month closed earlier does not rewrite when that happened. The response lists only the months this call actually closed.\n\nAfterwards, any attempt to create, change, or delete a plan or an expense in those months is rejected with `423` and the code `PERIOD_LOCKED`, including rows inside a CSV import. Moving a record out of a closed month is rejected too, because that changes the closed month's totals.",
        "operationId": "PeriodLocksController_lock_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePeriodLockDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The months this call closed. Already closed months are omitted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Neither months nor a quarter was supplied, or a month was malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "The access token is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close periods",
        "tags": [
          "Period locks"
        ]
      }
    },
    "/api/v1/period-locks/{month}": {
      "delete": {
        "description": "Reopens a single month, making its plans and expenses editable again.\n\nReopening is recorded in the audit log, with the month and the request that did it, so a period that was closed and then reopened leaves a trail rather than looking as though it was never closed.\n\nA month that is not closed answers 404.",
        "operationId": "PeriodLocksController_unlock_v1",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "path",
            "description": "The month to reopen, as YYYY-MM.",
            "schema": {
              "example": "2026-01",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The period is open."
          },
          "400": {
            "description": "The month is malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "The access token is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "That month is not closed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reopen a period",
        "tags": [
          "Period locks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Plan vs Actual API",
    "description": "Monthly spending targets, logged expenses, and variance reporting with locked periods.\n\n**Money** is an integer count of minor units. A field ending in `Minor` holds cents.\n\n**Months** are the string `YYYY-MM`.\n\n**Variance** is `actual - plan`. Variance percent is `null` when the plan is zero, never `NaN`.\n\n**Missing actuals** default to `0`. Pass `missingActuals=null` to receive `null` instead. Every report row carries `hasActual`, so a logged zero is never confused with nothing logged.\n\n**Locked periods** reject writes with `423` and the code `PERIOD_LOCKED`.\n\n**Deletes are soft.** A `DELETE` answers `204` and the record stops appearing in reads.\n\n**Sessions** travel in the `Authorization` header, never a cookie, so a mobile or desktop client is a first class caller. Access tokens are RS256 signed and short lived; refresh tokens are opaque, single use, and rotate.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorResponseDTO": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status, repeated in the body.",
            "example": 423
          },
          "code": {
            "type": "string",
            "description": "Stable code a client branches on. Never parse the message.",
            "enum": [
              "VALIDATION_FAILED",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "IMPORT_VALIDATION_FAILED",
              "PERIOD_LOCKED",
              "RATE_LIMITED",
              "INTERNAL"
            ],
            "example": "PERIOD_LOCKED"
          },
          "message": {
            "type": "string",
            "description": "What went wrong, written for a person.",
            "example": "2026-01 is locked and cannot be edited."
          },
          "details": {
            "type": "object",
            "description": "Machine readable specifics: the offending month, the failing rows, the invalid fields.",
            "example": {
              "month": "2026-01"
            }
          },
          "requestId": {
            "type": "string",
            "description": "Correlates this response with its log lines and the x-request-id header.",
            "example": "0f9c1e2a-7b3d-4c5e-9a8f-1b2c3d4e5f6a"
          },
          "path": {
            "type": "string",
            "description": "The path that produced the error.",
            "example": "/api/v1/plans"
          },
          "timestamp": {
            "type": "string",
            "description": "When the error was produced.",
            "example": "2026-01-15T10:04:11.212Z",
            "format": "date-time"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "requestId",
          "path",
          "timestamp"
        ]
      },
...
      "PeriodLockResponseDTO": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "description": "The closed month.",
            "example": "2026-01"
          },
          "lockedAt": {
            "type": "string",
            "description": "When it was closed.",
            "example": "2026-02-01T09:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "month",
          "lockedAt"
        ]
      },
      "CreatePeriodLockDTO": {
        "type": "object",
        "properties": {
          "months": {
            "description": "Months to close, each as YYYY-MM. Locking a month that is already closed is harmless.",
            "example": [
              "2026-01",
              "2026-02"
            ],
            "maxItems": 15,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "quarter": {
            "type": "string",
            "description": "A calendar quarter to close, which expands to its three months. Q1 is January through March.",
            "example": "2026-Q1"
          }
        }
      },
...
    }
  }
}